test distfitters
